# web/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "version": "1.0.0",
        "title": "Swipe Shopify"
    },
    "servers": [
        {
            "url": "/"
        }
    ],
    "components": {
        "securitySchemes": {
            "apiKeyAuth": {
                "type": "apiKey",
                "in": "header",
                "name": "x-access-token"
            }
        }
    },
    "security": [
        {
            "apiKeyAuth": []
        }
    ],
    "paths": {
...
        "/v1/statement/merchant/:merchantId": {
            "get": {
                "tags": ["Statement"],
                "summary": "Merchant List API",
                "produces": ["application/json"],
                "consumes": ["application/json"],
                "responses": {
                    "200": {
                        "description": "Data found."
                    },
                    "400": {
                        "description": "Data did not found."
                    }
                }
            }
        },
        "/v1/statement/preview/": {
            "get": {
                "tags": ["Statement"],
                "summary": "Statement Preview API",
                "produces": ["application/json"],
                "consumes": ["application/json"],
                "responses": {
                    "200": {
                        "description": "Data found."
                    },
                    "400": {
                        "description": "Data did not found."
                    }
                }
            }
        },
        "/v1/statement/list/": {
            "get": {
                "tags": ["Statement"],
                "summary": "Admin State List API",
                "produces": ["application/json"],
                "consumes": ["application/json"],
                "responses": {
                    "200": {
                        "description": "Data found."
                    },
                    "400": {
                        "description": "Data did not found."
                    }
                }
            }
        },
...
    },
...
}